requirements.txt
streamlit
# PyPDF2
google_generativeai
python-dotenv
requests